% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_types.r
\name{Enum}
\alias{Enum}
\title{Enum
Enum assertion}
\usage{
Enum(supplied_enum = NULL, null_ok = FALSE, ...)
}
\arguments{
\item{supplied_enum}{the enum to compare values against}

\item{null_ok}{whether null values are accepted}

\item{...}{the value to check assertions with}
}
\value{
any
}
\description{
Enum
Enum assertion
}
