% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_types.r
\name{Enum}
\alias{Enum}
\title{Enum assertion}
\usage{
Enum(supplied_enum = NULL, null_ok = FALSE, ...)
}
\arguments{
\item{supplied_enum}{an enum to compare values against}

\item{null_ok}{whether null values are accepted}

\item{...}{a value to check type assertions with}
}
\value{
any
}
\description{
Assertion for controlling input and return types.
}
\examples{
# If you leverage the {typed} package, you can make use
# of the `Enum()` assertion for type checking:

library(typed)
my_enum <- enum(a, b, c)

# Variable 'x' must correspond to an enum
Enum() ? x <- my_enum

# Variable 'x' must correspond to a variable in an enum
Enum(my_enum) ? x <- 2

# Argument 'x' must correspond to a variable in an enum
my_function <- ? function(x = ? Enum(my_enum)) {
    print("success!")
}

# Function must return an enum
func_return <- Enum() ? function() {
    return(
        enum(a, b, c)
    )
}
}
