% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumr-package.r
\docType{package}
\name{enumr-package}
\alias{enumr}
\alias{enumr-package}
\title{enumr: Enumerations for R}
\description{
Enums are lists that have unique key/pair values (called 'members'),
and cannot be modified after their definition. The main methods exported
are:
}
\details{
\itemize{
\item \code{\link[=enum]{enum()}} (object definition)
\item \code{\link[=as_enum]{as_enum()}} (object coercion)
\item \code{\link[=new_generic_enum]{new_generic_enum()}} (constructor)
\item \code{\link[=new_numeric_enum]{new_numeric_enum()}} (constructor)
\item \link[=enum]{Enum()} (type assertion)
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ElianHugh/enumr}
  \item Report bugs at \url{https://github.com/ElianHugh/enumr/issues}
}

}
\author{
\strong{Maintainer}: Elian Thiele-Evans \email{elianhte@gmail.com} (\href{https://orcid.org/0000-0001-8008-3165}{ORCID})

}
\keyword{internal}
