% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumr-package.r
\docType{package}
\name{enumr}
\alias{enumr-package}
\title{enumr: Enumerations for R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Implement static enumerations for R.
}
\details{
The \strong{enumr} package implements enumerations (or 'enums').
Enums are lists that have unique key/pair values (called 'members'),
and cannot be modified after their definition.

General Resources:
\itemize{
\item Enumr's enums are heavily inspired by the typescript
data structure of the same name.
See \href{https://www.typescriptlang.org/docs/handbook/enums.html}{here}
for how enums function in typescript
\item Enumr's \href{https://elianhugh.github.io/enumr/index.html}{pkgdown site}
}

Main exported methods:
\itemize{
\item making enums: \code{\link[=enum]{enum()}}, \code{\link[=as_enum]{as_enum()}}, \code{\link[=new_generic_enum]{new_generic_enum()}},
\code{\link[=new_numeric_enum]{new_numeric_enum()}}
\item type checking: \code{\link[=Enum]{Enum()}}, \code{\link[=is_enum]{is_enum()}}
\item printing and presenting enums: \code{\link[=print.enum]{print.enum()}}, \code{\link[=format.enum]{format.enum()}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ElianHugh/enumr}
  \item Report bugs at \url{https://github.com/ElianHugh/enumr/issues}
}

}
\author{
\strong{Maintainer}: Elian Thiele-Evans \email{elianhte@gmail.com} (\href{https://orcid.org/0000-0001-8008-3165}{ORCID})

}
