% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumRsimulation.R
\name{enumRsimulation}
\alias{enumRsimulation}
\title{enumR simulation function}
\usage{
enumRsimulation(seed = 12345, ndatasets = 200, patternmatrix = NULL,
  effectmatrix = NULL, nfactors = 5, loading = 0.5, items = NULL,
  items_p_f = 5, itemsR_p_f = 2, itemsR = NULL, loading_norm = FALSE,
  loading_norm_sd = 0.05, rfactors = 0, r_norm = FALSE,
  r_norm_SD = 0.015, samplesize = 300, method = "ml",
  rotation = "oblimin", custom_item = FALSE, f1_items = NULL,
  f2_items = NULL, f3_items = NULL, f4_items = NULL, f5_items = NULL,
  f6_items = NULL, f7_items = NULL, f8_items = NULL, f9_items = NULL,
  f10_items = NULL, f1_itemsR = NULL, f2_itemsR = NULL,
  f3_itemsR = NULL, f4_itemsR = NULL, f5_itemsR = NULL,
  f6_itemsR = NULL, f7_itemsR = NULL, f8_itemsR = NULL,
  f9_itemsR = NULL, f10_itemsR = NULL)
}
\arguments{
\item{seed}{numeric value for setting a seed. Allows results to be reproduced.}

\item{ndatasets}{Number of datasets generated}

\item{patternmatrix}{OPTIONAL matrix of factor loadings. If not provided, one is generated from \code{rfactors} and \code{nfactors}.}

\item{effectmatrix}{OPTIONAL matrix of item loadings by factor. If not provided, one is generated from \code{items} and \code{nfactors}.}

\item{nfactors}{Number of factors to extract.}

\item{loading}{Factor loading magnitude. Default is .5}

\item{items_p_f}{number of items per factor}

\item{itemsR_p_f}{number of reverse-scored items per factor}

\item{itemsR}{Number of items per factor that are reverse scored}

\item{loading_norm}{If TRUE, generate factor loadings that average to \code{loading}}

\item{loading_norm_sd}{If \code{loading_norm} is TRUE, standard deviation of loadings.}

\item{rfactors}{Intercorrelations between factors. Default is 0}

\item{r_norm}{If TRUE, generate normally distributed factor loadings with  mean \code{loading} and sd \code{r_norm_sd}. Does not check if correlation matrix is positive definite.}

\item{samplesize}{Number of observations}

\item{method}{factoring method – fm="pa" Principal Axis Factor Analysis, fm = "minres" minimum residual (OLS) factoring fm="ml" Maximum Likelihood FA, fm="pc" Principal Components"}

\item{rotation}{what rotation to use c("none", "varimax", "oblimin","promax")}

\item{custom_item}{If TRUE, overrides \code{items_p_f} and \code{itemsR_p_f}. Primary used for importing from shiny interface.
@param items OPTIONAL vector of total number of items per factor}

\item{f1_items}{Total number of items on factor 1}

\item{f2_items}{Total number of items on factor 2}

\item{f3_items}{Total number of items on factor 3}

\item{f4_items}{Total number of items on factor 4}

\item{f5_items}{Total number of items on factor 5}

\item{f6_items}{Total number of items on factor 6}

\item{f7_items}{Total number of items on factor 7}

\item{f8_items}{Total number of items on factor 8}

\item{f9_items}{Total number of items on factor 9}

\item{f10_items}{Total number of items on factor 10}

\item{f1_itemsR}{Number of reverse-scored items on factor 1}

\item{f2_itemsR}{Number of reverse-scored items on factor 2}

\item{f3_itemsR}{Number of reverse-scored items on factor 3}

\item{f4_itemsR}{Number of reverse-scored items on factor 4}

\item{f5_itemsR}{Number of reverse-scored items on factor 5}

\item{f6_itemsR}{Number of reverse-scored items on factor 6}

\item{f7_itemsR}{Number of reverse-scored items on factor 7}

\item{f8_itemsR}{Number of reverse-scored items on factor 8}

\item{f9_itemsR}{Number of reverse-scored items on factor 9}

\item{f10_itemsR}{Number of reverse-scored items on factor 10
...    parameters to pass to the factor analysis program The most important of these is if using a correlation matrix is covmat= xx}

\item{cor}{What kind of correlation to find, defaults to covariance matrix, but see fa for the choices}

\item{r_norm_sd}{If \code{r_norm} is TRUE, standard deviation for normally distributed intercorrelations.}

\item{true_model}{If TRUE, Estimates model using \code{nfactors} only. Is used in \code{enumRsimulation} function. Default is FALSE.}

\item{use}{If doing covariances or Pearson R, should we use "pairwise" or "complete cases"}

\item{items_p_f}{number of items per factor}

\item{itemsR_p_f}{number of reverse-scored items perfactor}

\item{custom_item}{If TRUE, overrides \code{items_p_f} and \code{itemsR_p_f}. Primary used for importing from shiny interface.}
}
\value{
dataframe
}
\description{
This function simulates factor analyses based on numerous parameters, listed below.
}
