% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.r
\name{$<-.enum}
\alias{$<-.enum}
\alias{[<-.enum}
\alias{[[<-.enum}
\alias{[.enum}
\alias{[[.enum}
\alias{$.enum}
\title{Extract Parts of Enum}
\usage{
\method{$}{enum}(x, arg) <- value

\method{[}{enum}(x, arg) <- value

\method{[[}{enum}(x, arg) <- value

\method{[}{enum}(x, arg)

\method{[[}{enum}(x, arg)

\method{$}{enum}(x, arg)
}
\arguments{
\item{x}{object to examine}

\item{arg}{indices to extract}

\item{value}{value}
}
\description{
Access parts of enum through numeric or character indices
}
\section{Functions}{
\itemize{
\item \code{$<-.enum}: Enum

\item \code{[<-.enum}: Enum

\item \code{[[<-.enum}: Enum

\item \code{[.enum}: Enum

\item \code{[[.enum}: Enum

\item \code{$.enum}: Enum
}}

