% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenData.R
\name{GenData}
\alias{GenData}
\title{Generate Data from factor loadings}
\usage{
GenData(fmodel, effect, samplesize, inames = NULL)
}
\arguments{
\item{fmodel}{Matrix of factor loadings for all items}

\item{effect}{effects matrix}

\item{samplesize}{number of observations generated}

\item{names}{names of variables}
}
\value{
data.frame of observed scores
}
\description{
\code{GenData} returns a data.frame of observed scores based on matrix of factor loadings, effects matrix. The heart of this function was adapted from \url{http://personality-project.org/r/r.datageneration.html}
}
\details{

}
