% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenFactorMatrix.R
\name{GenFactorMatrix}
\alias{GenFactorMatrix}
\title{Generate Factor Loadings Matrix}
\usage{
GenFactorMatrix(nfactors = 5, items = c(5, 5, 5, 5, 5), itemsR = c(2, 2,
  2, 2, 2), loading = 0.5, loading_norm = FALSE, loading_norm_sd = 0.025)
}
\arguments{
\item{nfactors}{Number of factors}

\item{items}{vector of total number of items per factor}

\item{itemsR}{Number of items per factor that are reverse scored}

\item{loading}{Factor loading magnitude. Default is .5}

\item{loading_norm}{If TRUE, generate factor loadings that average to \code{loading}}

\item{loading_norm_sd}{If \code{loading_norm} is TRUE, standard deviation of loadings.}
}
\value{
\code{matrix} of n factors by i items
}
\description{
\code{GenFactorMatrix} returns a matrix of factor loadings, based on number of items, factors, and loadings
}
