% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_enum.r
\name{new_numeric_enum}
\alias{new_numeric_enum}
\alias{new_generic_enum}
\title{Enum Constructors}
\usage{
new_numeric_enum(.enum_data)

new_generic_enum(.enum_data)
}
\arguments{
\item{.enum_data}{named list of arguments}
}
\description{
\code{new_numeric_enum()} creates and validates a
numeric enum from a named list of arguments

\code{new_generic_enum()} creates and validates a
generic enum from a named list of arguments
}
\examples{
new_numeric_enum(list(x = 5))
new_generic_enum(list(x = "str", y = mtcars))
}
\seealso{
\code{\link[=enum]{enum()}}
}
