% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_enum.r
\name{only_values_supplied}
\alias{only_values_supplied}
\title{Check that the enum has atomic members without names
Returns TRUE if this is the case, which will cause
enumr to shoot an error}
\usage{
only_values_supplied(obj)
}
\arguments{
\item{obj}{an enum to check}
}
\description{
Check that the enum has atomic members without names
Returns TRUE if this is the case, which will cause
enumr to shoot an error
}
