% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print.enum}
\alias{print.enum}
\alias{format.enum}
\title{Print object}
\usage{
\method{print}{enum}(x, ...)

\method{format}{enum}(x, ...)
}
\arguments{
\item{x}{an object to print}

\item{...}{further arguments passed to as.list}
}
\description{
Print object

Format string
}
\section{Functions}{
\itemize{
\item \code{print.enum}: Object

\item \code{format.enum}: Object
}}

