% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.r
\name{length.enum}
\alias{length.enum}
\alias{names.enum}
\alias{as.list.enum}
\title{Length}
\usage{
\method{length}{enum}(x)

\method{names}{enum}(x)

\method{as.list}{enum}(x, all.names = FALSE, sorted = FALSE, ...)
}
\arguments{
\item{x}{object to coerce to list}

\item{all.names}{copy all values or only those that do not begin with a dot}

\item{sorted}{should the list be sorted by names?}

\item{...}{other args}
}
\description{
Length

Names

As List
}
\section{Functions}{
\itemize{
\item \code{length.enum}: Methods

\item \code{names.enum}: Methods

\item \code{as.list.enum}: Methods
}}

