% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.r
\name{str.enum}
\alias{str.enum}
\title{Display an enum's internal structure}
\usage{
\method{str}{enum}(object, ..., nest.lev = 0L)
}
\arguments{
\item{object}{an enum to examine}

\item{...}{dots to pass to further methods}

\item{nest.lev}{the nest level that str should start at}
}
\description{
Display the internal structure of an enum,
listing enum members and the variables defined
within them.
}
