% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_enum.r
\name{validate_enum_definition}
\alias{validate_enum_definition}
\alias{validate_numeric_enum}
\alias{validate_generic_enum}
\title{Enum Validator}
\usage{
validate_enum_definition(.enum_data)

validate_numeric_enum(.enum_data)

validate_generic_enum(.enum_data)
}
\arguments{
\item{.enum_data}{named list of arguments}
}
\description{
\code{validate_enum_definition()} checks the enum arguments
for validity

\code{validate_numeric_enum()} checks named list of arguments
for validity

\code{validate_generic_enum()} checks generic enums
for validity
}
\section{Validation}{

\code{validate_enum_definition()} checks that the names
and values supplied to the constructor are unique,
and ensures that all values supplied have names.


\code{validate_numeric_enum()} checks that, when evaluated,
all the enum values are unique
and can be interpreted as numeric


\code{validate_generic_enum()} check that all
arguments supplied to the constructor
have a name and value.
}

