% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.r
\name{enum}
\alias{enum}
\title{Create a new enumeration}
\usage{
enum(..., .subtract_value = 0)
}
\arguments{
\item{...}{list of enumeration arguments}
}
\value{
An enumeration (enum), a list of unique name/value pairs.
}
\description{
\code{enum()} creates a new enumeration, and infers the the enumeration type
based on the arguments supplied to it. Enumerations are similar to lists,
but:
\itemize{
\item cannot be assigned to after definition
\item all values and names must be unique
}

There are two types of enumeration: generic and numeric. The main difference
between the two is that generic enumerations must have all their values
explicitly defined. Numeric enums, by contrast, infer the values of
their members, and must have only numeric values.

Numeric enum values are inferred when a member does not have a value assigned
to it. The value of the member is then calculated as +1 to the last
member's value, or the current member's index. Incrementation for values is
prioritised over index as value.
}
\details{
Under the surface, enums are actually lists contained within
locked environments. This is so that the enum bindings cannot be modified,
and the enum order is maintained. S3 methods are defined for
\code{$}, \code{[}, and \code{[[}, which access the enum list directly.
}
\examples{
# Generic Enum
enum(apple = "apple", pear = "pear")

enum(dat1 = mtcars, dat2 = iris, dat3 = PlantGrowth)

# Numeric Enum
enum(style, warning, error)

enum(a = 50, b = .$a * 2)
}
\seealso{
\code{\link[=new_numeric_enum]{new_numeric_enum()}}, \code{\link[=new_generic_enum]{new_generic_enum()}}, \code{\link[=as_enum]{as_enum()}}
}
