% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumr-package.r
\docType{package}
\name{enumr-package}
\alias{enumr}
\alias{enumr-package}
\title{enumr: Enumerations for R}
\description{
Enums are lists that have unique key/pair values (called 'members'),
and cannot be modified after their definition. \code{\link[=enum]{enum()}} (object definition)
and \link[=enum]{Enum()} (type assertion) are the main methods contributed by enumr.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ElianHugh/enumr}
  \item Report bugs at \url{https://github.com/ElianHugh/enumr/issues}
}

}
\author{
\strong{Maintainer}: Elian Thiele-Evans \email{elianhte@gmail.com} (\href{https://orcid.org/0000-0001-8008-3165}{ORCID})

}
\keyword{internal}
