% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumr-package.r
\docType{package}
\name{enumr}
\alias{enumr}
\alias{enumr-package}
\title{enumr: Enumerations for R}
\description{
Enums are lists that have unique key/pair values (called 'members'), and cannot be modified after their definition.
\code{\link[=enum]{enum()}} (object definition) and \link[=enum]{Enum()} (type assertion) are the main methods contributed by enumr.
}
\author{
\strong{Maintainer}: Elian Thiele-Evans \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
