% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumR.R
\name{enumR}
\alias{enumR}
\title{enumR function}
\usage{
enumR(x, nfactors = 8, rotation = "oblimin", diagonal = FALSE,
  method = "ml", samplesize = NULL, use = "pairwise", cor = "cov",
  truemodel = FALSE, ...)
}
\arguments{
\item{x}{a correlation matrix or a data matrix
...    parameters to pass to the factor analysis program The most important of these is if using a correlation matrix is covmat= xx}

\item{nfactors}{Maximum number of factors to extract. Needs to be larger than expected.}

\item{rotation}{what rotation to use c("none", "varimax", "oblimin","promax")}

\item{diagonal}{Should we fit the diagonal as well}

\item{method}{factoring method – fm="pa" Principal Axis Factor Analysis, fm = "minres" minimum residual (OLS) factoring fm="ml" Maximum Likelihood FA, fm="pc" Principal Components"}

\item{samplesize}{Number of observations if doing a factor analysis of correlation matrix. This value is ignored by VSS but is necessary for the ML factor analysis packag}

\item{use}{If doing covariances or Pearson R, should we use "pairwise" or "complete cases"}

\item{cor}{What kind of correlation to find, defaults to covariance matrix, but see fa for the choices}

\item{true_model}{If TRUE, Estimates model using \code{nfactors} only. Is used in \code{enumRsimulation} function. Default is FALSE.}
}
\value{
dataframe
}
\description{
The enumR function is a heavily modified version of the \code{vss} function from \code{Psych} Package]. Both functions are designed to help researchers determine the number of factors in their exploratory factor analysis. This function includes a variarity of methods.
}
