% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_types.r
\name{is_enum}
\alias{is_enum}
\title{Test if the object is an enum}
\usage{
is_enum(x)
}
\arguments{
\item{x}{value to check if is enum}
}
\value{
TRUE if the object inherits from the
enum class, and FALSE if it does not. Does not
distinguish between numeric or generic enums.
}
\description{
Test if the object is an enum
}
