% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{printing}
\alias{printing}
\alias{print.enum}
\title{Printing enums}
\usage{
\method{print}{enum}(x, ...)
}
\arguments{
\item{x}{enum to print}

\item{...}{values passed to further methods}
}
\description{
Pretty print enums with the print.enum method
}
