% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.r
\name{S3-Methods}
\alias{S3-Methods}
\alias{str.enum}
\alias{length.enum}
\alias{names.enum}
\alias{as.list.enum}
\alias{as.character.enum}
\alias{as.complex.enum}
\title{General S3 Methods for Enums}
\usage{
\method{str}{enum}(object, ...)

\method{length}{enum}(x)

\method{names}{enum}(x)

\method{as.list}{enum}(x, ...)

\method{as.character}{enum}(x, ...)

\method{as.complex}{enum}(x, ...)
}
\arguments{
\item{object}{enum}

\item{...}{arguments to pass to as.list}

\item{x}{enum}
}
\description{
Various S3 methods that improve interacting with
enum data types
}
\details{
These are some details
}
