% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.r
\name{S3-Methods}
\alias{S3-Methods}
\alias{length.enum}
\alias{names.enum}
\alias{as.list.enum}
\alias{as.character.enum}
\title{General S3 Methods for Enums}
\usage{
\method{length}{enum}(x)

\method{names}{enum}(x)

\method{as.list}{enum}(x, ...)

\method{as.character}{enum}(x, ...)
}
\arguments{
\item{x}{enum}

\item{...}{arguments to pass to as.list}
}
\description{
Various S3 methods that improve interacting with
enum data types.

These rely on coercing enums to list types before
applying any methods to the enum.
}
