% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.r
\name{subsetting}
\alias{subsetting}
\alias{[.enum}
\alias{[[.enum}
\alias{$.enum}
\alias{$<-.enum}
\alias{[<-.enum}
\alias{[[<-.enum}
\title{Subsetting enums}
\usage{
\method{[}{enum}(x, arg)

\method{[[}{enum}(x, arg)

\method{$}{enum}(x, arg)

\method{$}{enum}(x, arg) <- value

\method{[}{enum}(x, arg) <- value

\method{[[}{enum}(x, arg) <- value
}
\arguments{
\item{x}{enum to subset}

\item{arg}{value to subset enum with, generally the indices of the enum}

\item{value}{assigning values to enums is not allowed}
}
\description{
Access enum members through subsetting with numeric or character indices
}
\details{
Assigning to enum members is not supported, and will
throw an error if attempted.
}
