% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.r
\name{subsetting}
\alias{subsetting}
\alias{[.enum}
\alias{[[.enum}
\alias{$.enum}
\alias{$<-.enum}
\alias{[<-.enum}
\alias{[[<-.enum}
\title{Subsetting enums}
\usage{
\method{[}{enum}(x, arg)

\method{[[}{enum}(x, arg)

\method{$}{enum}(x, arg)

\method{$}{enum}(x, arg) <- value

\method{[}{enum}(x, arg) <- value

\method{[[}{enum}(x, arg) <- value
}
\arguments{
\item{x}{enum to subset}

\item{arg}{value to subset enum with, generally the indices of the enum}

\item{value}{assigning values to enums is not allowed}
}
\description{
Access enum members through subsetting with numeric or character indices.

Enums are converted to a list type through the
\code{\link{as.list}}][base::as.list()] method, before being subset by an index.
}
\details{
Enums are not intended to be used as interactive data structures.
Assigning to enum members is not supported, and will
throw an error if attempted.

If you require an enum to be modifiable, consider using a
\code{\link[base:list]{list}} instead.
}
\examples{
a <- enum(a,b,c)
a[1]
a[[1]]
a$a
}
