% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_enum.r
\name{validate_enum_definition}
\alias{validate_enum_definition}
\alias{validate_numeric_enum}
\alias{validate_generic_enum}
\title{Enum Validator}
\usage{
validate_enum_definition(.enum_data)

validate_numeric_enum(.enum_data)

validate_generic_enum(.enum_data)
}
\arguments{
\item{.enum_data}{named list of arguments}
}
\value{
input data

input data

input data
}
\description{
The basis for all enum definitions, \code{validate_enum_definition()}
checks the enum arguments for internal consistency

\code{validate_numeric_enum()} checks numeric enum arguments
for internal consistency with the numeric enum class

\code{validate_generic_enum()} checks generic enums
for internal consistency with the generic enum class
}
\section{Validation}{

\code{validate_enum_definition()} checks that all values supplied have names.


\code{validate_numeric_enum()} checks that, when evaluated,
all the enum members are unique,
and can be interpreted as numeric


\code{validate_generic_enum()} check that all
arguments supplied to the constructor
have a unique name and value.
}

